+++ Source/ACE/Command/Handlers/DebugCommands.cs
@@ -432,6 +432,8 @@ namespace ACE.Command.Handlers
             Console.WriteLine($"Error locating the player.");
         }
 
+        // TODO: Replace later with a command to spawn a generator at the player's location
+        /*
         /// <summary>
         /// Debug command to spawn a creature in front of the player and save it as a static spawn if the static option is specified.
         /// </summary>
@@ -480,6 +482,7 @@ namespace ACE.Command.Handlers
                     ChatMessageType.Broadcast);
             }
         }
+        */
 
         /// <summary>
         /// Debug command to kill a targeted creature so it drops a corpse.
@@ -606,7 +609,6 @@ namespace ACE.Command.Handlers
             session.Network.EnqueueSend(positionMessage);
         }
 
-        /*
         /// <summary>
         /// Debug command to test the ObjDescEvent message. 
         /// </summary>
@@ -647,7 +649,6 @@ namespace ACE.Command.Handlers
                 ChatPacket.SendServerMessage(session, "Please enter a value greater than 0x10000000 and less than 0x1000086C", ChatMessageType.Broadcast);
             }
         }
-        */
 
         /// <summary>
         /// Debug command to learn a spell.