+++ changelog.md
@@ -1,5 +1,15 @@
 # ACEmulator Change Log
 
+### 2017-06-25
+[Ripley]
+* Expanded upon @Lidefeath's generator work
+* Added complete generator test of tutorial dungeon setup for review
+
+### 2017-06-24
+[Lidefeath]
+* Remove old creature spawning and generator stuff
+* Implement creature spawning from generator objects, nested or hierarchical generators are still to do
+
 ### 2017-06-23
 [ddevec]
 * Fix to teleport freezing characters in place (Wasn't clearing requested location)